Comments and stuff
Added all the stuff. Shooter is still incomplete - it is in another
project and will be merged when it can be tested since we haven't used
pneumatics while programming in C++ yet. Every file should be properly
commented now, and once we have a robot to test on, it should be able to
drive.